Add twisted Cartesian powers; hard-core coverage 378 -> 416
Paper section 5.4. Given a magma satisfying E and two endomorphisms T, U,
the twisted operation x ◇' y := Tx ◇ Uy satisfies E again when T, U obey
the relations of the twisting semigroup Twist_E, and a Cartesian power M^k
supplies those as coordinate shifts.

The twisted magma has n^k elements and 22 laws use six variables, so a
sweep of the paper's own example would need 32^6 assignments. The check
decomposes by coordinate instead: eval(Op(l,r), i) = eval(l, s(i)) ◇
eval(r, t(i)), so a law reads at most six entries per root coordinate.
Cost is k·n^leaves per law rather than n^(k·vars).

Reproduces Example 5.11 exactly: F_2 under NAND, k=5, shifts (+1,-1)
witnesses E1485 ⊭ E151, which the paper says "does not seem to be easily
refuted by any of the other methods discussed". That one instance closes
all 38 hard-core pairs under E1485 and E2162 — the cluster no linear or
affine instance touched.

Two saturation results worth recording. Widening the modulus grid to
m <= 96 (354,247 linear instances) discharges no additional hard-core
pair. Adding 96,570 twisted powers over 3-element bases yields 3219 new
signatures and three points of whole-graph coverage (96.85% -> 99.85%),
and again zero additional hard-core pairs. Compute is not the binding
constraint; the set of construction kinds is.

24 tests, clippy clean.

+++ src/nc.rs
@@ -36,7 +36,7 @@ pub const fn word_of(index: usize) -> (u32, u32) {
 /// Degrees `(deg_a, deg_b)` of a word, for use in commutative rings.
 pub const fn word_degrees(index: usize) -> (u32, u32) {
     let (len, bits) = word_of(index);
-    let db = (bits as u32).count_ones();
+    let db = bits.count_ones();
     (len - db, db)
 }
